requirements.txt
streamlit
torch
torchvision
ultralytics
pillow
opencv-python
numpy
pytesseract
google-generativeai
langchain-google-genai
langchain-core-messages
langchain
pyttsx3